pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reasonflow"
version = "0.1.0"
description = "SDK-first, MCP-native agent orchestration framework"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "litellm>=1.40.0",
    "click>=8.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
all = ["reasonflow[mcp,dev]"]

[project.scripts]
reasonflow = "reasonflow.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/reasonflow"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
